Changed requires so that importing `rugged` and `pry` work for me.
I never got to the root of the problem, but when I first checked
out the project, it wasn't able to find the `rugged` package which
was installed via bundler. Installing a separate version via `gem`
wasn't able to use SSL. It seems that the version of `rugged`
built from git was able to use SSL, but wasn't being found by
`rubygems`, and a bit of searching led me to include the bundler
setup lines. It now works - but I can't explain why :-)